fix: Fix race that causes incorrect `global_last_seen_lsn` delivery on up-to-date, non-live repsonses (#2450)
Fixes https://github.com/electric-sql/electric/issues/2447


## Problem
- We were determining the `global_last_seen_lsn` at the end of the
request processing pipeline
- The `last_offset` available for the shape was determined at the start,
and that is what we use to decide if the response is up-to-date or not
- In between reading `last_offset` at the start (and determining
response was up-to-date), and reading `global_last_seen_lsn` at the end,
a change might have come in, increasing the global seen LSN and adding a
change to the current shape's log, which won't be seen by the response
as it will read up to and including the `last_offset` which was
determined at the start.


## Testing
- I first tried to reproduce the case with integration tests
(`test/electric/plug/router.exs`) and adding arbitrary `Process.sleep`
calls until I could reproduce the situation - but since this was a race
this was a bit hard to add as a proper test
- I moved to `test/electric/shapes/api_test.exs` where I could use our
mocks to simulate the global LSN moving right after the
`get_chunk_end_log_offset` call is made - this test could now
consistently detect this simulated race. Not as good as an integration
test but fully reproducible, controlled, and fast to run.


## Solution
- Moved the `global_last_seen_lsn` retrieval to right before we
determine the actual log `last_offset`, we prefer potentially being
veeery slightly far behind the actual global last seen LSN than having
the above race occur.
- Made this step into a `determine_global_last_seen_lsn` method to match
our existing structure, and placed it in all appropriate places
- Right _before_ we determine the chunk end offset when validating the
request
- Right before we re-determine the chunk end offset during a live
request when we receive new changes
- Right before we return an empty live response after a long poll
timeout to ensure that we get the latest global seen LSN