feat: remove full table info from shape definitions to lower mem (#2454)
## Reasons

We have a process setup where we have 2 layers of supervisors which
store entire child spec of their children as state, plus the
`ShapeLogCollector` that the shape consumers register with, plus the
consumers themselves -- all these processes happen to have shapes in
their state. _Arguably, we can inject shapes into consumers bypassing
supervisors, but that's a different PR._ These shapes, however, happened
to know entire schema of the table they were defined against, which was
duplicated a lot. This caused a lot of unnecessary memory usage when the
table schema was sufficiently large.

So, in numbers, for a 75 column table, without any transactions going
through (just shape creation)
Baseline, no shapes: 71kB
- Current, 300 shapes
	- Total (for shape-related processes): 64.1MB
	- Total overhead in shared processes: 28MB 
	- Total overhead in shared per shape: 93kB
	- Per-shape process total: 35.9MB
	- Per-shape 119kB
- With this PR, 300 shapes
	- Total (for shape-related processes): 21MB
	- Total overhead in shared processes: 7.3MB
	- Total overhead in shared per shape: 24kB
	- Per-shape process total: 13MB
	- Per-shape 45kB

So for a sufficiently large table, this PR drops used memory a lot, 60%
in this particular case. Further optimizations of state are possible and
we'll do them, but this is the lowest hanging fruit.

## Changes

Shapes now don't store their table's entire schema. That's OK when
processing transactions or querying the DB, but that's not enough to
invalidate the shape when the schema changes, and it can change while
Electric is down as well.

To address this, now the burden of figuring out whether a
schema-breaking change has happened from the incoming relations now lies
upon `ShapeLogCollector`. It tracks all seen relation messages (which
should be relatively rare unless PG is constantly disconnecting), and
marks the columns that have changed within this relation as compared to
a previously seen version. It also saves these last-seen relations when
all shapes have reacted to it and invalidated themselves.