docs: Document secure and insecure modes (#2405)
Docs update for https://github.com/electric-sql/electric/pull/2404

I think we should also extend other parts of the documentation to
explain that this secret must be passed when making shape requests.
Otherwise, people will follow the docs but it won't work. So either the
docs show usage in INSECURE mode which doesn't seem ideal, or we add the
secret where needed. But i'm a bit afraid of polluting code
snippets/docs with a secret. Open for suggestions on the best way to
communicate this in the docs.