fix: Add tests for binary columns (#2480)
Was triaging https://github.com/electric-sql/electric/issues/2469 and I
added two integration tests:

1. Ensure most types get replicated correctly, both for the initial
snapshot and subsequent log updates
2. Ensure large binaries (>1 MB) get replicated correctly as well

Both tests are passing but I think they are worth including in the test
suite.

I think we need to change the title of the above issue as looking at it
more closely the problem is that we return a body with a 204 response,
which is against the spec (and is what the error is reporting as well).