fix: Return 409 must-refetch instead of 400 on shape definition and handle mismatch (#2476)
Fixes https://github.com/electric-sql/electric/issues/2474

Currently, Electric returns a 400 error code when the shape definition
and shape handle do not match.
However, we discovered a pattern for handling dependent shapes on the
proxy where shape B is constructed based on the results for shape A. For
example, if we want to load all teams for a user we would first create a
shape to get all team_ids a user belongs to and then another shape that
selects all teams where `team_id IN (team_id_1, ..., team_id_n)`. The
definition of the second shape changes whenever the results of the first
shape change. For this to work properly, we need Electric to return 409
must-refetch with the handle of the new shape when the shape definition
changes.