fix: Synchronously refresh publication and clear cache if failed (#2490)
Followup to https://github.com/electric-sql/electric/pull/2487

In the above PR I fixed an issue that would cause the publication
manager to exit, which made Electric go into a restart loop.

There were two issues that compounded that problem:
1. I had previously converted the recovery of the shapes on the
publication manager asynchronous, using `cast` instead of `call`, which
means that Electric went ahead with initialisation and marked itself
healthy despite not managing to actually recover - if it had waited for
the recovery to finish and fail the container would never have been
marked as healthy and never deployed.
2. If a recovery fails, for whatever reason, it might enter into an
infinite recovery loop as it is likely some of the persisted state from
which we are trying to recover is causing the issue. We should be able
to recover from that since we consider the storage/cache disposable.

In order to solve the above issues, I implement the following:
1. Convert the calls back to synchronous calls, increasing the timeout
for `ShapeCache` initialisation slightly, as that was the original
purpose of making them asynchronous (it might take a while to recover
things).
a. I'm not entirely sure if this is a bit of an anti-pattern - we could
place the initialisation in a `{:continue, :restore_shapes}` after the
init - my concern is that if the rest of Electric continues on and
initialises everything else before the restore is done we might have a
problem.
2. Use `try..catch..after` blocks to ensure cleanup occurs if any
recovery operation fails, such that Electric will still manage to start
after the supervisor restarts the process and the old, problematic state
is cleaned up.
a. It's a bit of a brutal approach, I've tested it by tweaking the
publication manager to fail every now and then and it works really
nicely, where it will erase the stored shapes and restart and continues
on like normal.
b. We could theoretically place this at a higher level, such as the
`start_link` of the `ShapeCache`, so that we always clear the cache if
initialisation fails, but I wanted to keep the scope tight and only to
things that we have observed during real operation.


I can spend time to add a sort of integration test with a failing
`PublicationManager` to better simulate this situation in a test suite
rather than ad hoc.

Tagging @balegas as well as we discussed this change