fix: Catch and surface body consumption errors in regular flow (#2500)
Closes https://github.com/electric-sql/electric/issues/2497

- Wrap fetch such that we ensure that we can not only complete the
request but also consume the fetch body in full - and any errors arising
from that are still surfaced as `FetchError`
- Always throw errors from our fetch even if the status code is not in
the `>= 400 && < 500` range - if we don't, it fails in the next step as
`response` is undefined.

In this way, errors during the consumption of the body stream are
handled the same as other network errors, interrupting the stream,
notifying subscribers, etc.