feat: Do not cache >= 400 repsonses and non-GET methods (#2501)
We currently cache _everything_, including 400 responses which can lead
to bad behaviours in long delays in recovering.

This PR sets an explicit `no-cache` header for anything but `GET` and
`OPTIONS` requests, and any response that is `>= 400`.

It might be worth discussing what to do with 409 responses - perhaps we
actually might want to cache something but for a very short time.