fix: disallow generated columns in shapes (#2507)
Closes #2448 

Generated columns cannot be included in a publication by PostgreSQL, so
we disallow them in shapes. When doing `CREATE PUBLICATION p FOR TABLE
t` without a column list, generated columns are omitted. When doing an
explicit `FOR TABLE t (id)` (if `id` is generated), then it says `ERROR:
cannot use generated column "id" in publication column list`.

We could, later, allow generated columns based on the subset of where
clause functions we support by filling the stored values on our side
too, but that's a way off.