feat(sync-service): Add LRU shape expiry (#2514)
With this PR, once the number of shapes exceeds `max_shapes` shapes are
expired, least recently used first.

`max_shapes` is currently fixed and specified with the new
`ELECTRIC_EXPERIMENTAL_MAX_SHAPES` environment variable. The env var has
`EXPERIMENTAL` in it's name to convey that we will not be supporting
this variable in the future since we would prefer that `max_shapes` was
dynamically calculated. However for trigger.dev we can see setting
max_shapes to 8000 should prevent WAL lag and shapes that have not been
used for about two hours are expired which suits their needs.

A shapes `last_read` time is not currently persisted to disk. That has
been left to a later PR. For trigger.dev this is fine since they do not
currently recover shapes on restart. In the event that shape recovery
was used with this implementation of cache expiry, shapes not used since
restoration would have a `last_read` time of the time of recovery and so
maybe be expired in an unexpected order.