fix!: Return 200 for empty live response (#2486)
Closes https://github.com/electric-sql/electric/issues/2469

Need to address whether this is a breaking change, if we want to do it
experimentally with a flag, whether we instead want to keep 204
responses and move info to headers.

Opening this PR so we can have the scope of the changes ready.

BREAKING CHANGE: community clients relying on `204` alone for up-to-date
logic might break - live responses now always return a `200` with a
body/