fix: Maintain backwards compatibility with 204s (#2532)
Closes https://github.com/electric-sql/electric/issues/2538

Maintain backwards compatibility after
https://github.com/electric-sql/electric/pull/2486

The change was forwards compatible (old clients work with new servers)
but not backwards compatible (new clients didn't work with old servers -
old enough that they wouldn't return a body with the 204).

NOTE: The change was actually both forwards and backwards compatible as
of the official 1.0.0 release, as we were already (wrongly) including a
body with 204 responses at that point. I think that we should not
reinsert this for the sake of backwards compatibility for versions prior
to the official 1.0.0.