fix: Handle db does not exist error as a fatal error (#2561)
This PR handles the Postgres "database does not exist" error which can
occur when a user deletes their database. This is a fatal error we can't
recover from so we now also handle it like that. I modified the shutdown
method to log a warning rather than an error since the shutdown is not
unexpected since we are shutting it down on purpose as a response to the
DB no longer existing. This will also avoid that this error pops up in
Sentry for our cloud.

Fixes https://github.com/electric-sql/electric/issues/2552

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>