Remove redundant call to LsnTracker.reset()
First of all, the ETS table which LsnTracker wants to update does not
exist yet at this point in the execution.

Secondly, the call is not needed because when the timeline change is
detected, the shape cache purges all shapes and sets its internal
last_processed_lsn value to 0 which is the same value is initializes the
LsnTracker table with.