sync-service: Set shape cache restore timeout per-shape, not globally (#2575)
Currently we have a hard-coded timeout of 20s for `ShapeCache.init/1` to
complete. Since we're starting all existing shape consumers within this
callback 20s may not be long enough if there are enough shapes.

Rather than limit the duration of the entire shape recovery process,
this PR wraps each shape-restart in a timeout (currently 5s) and leaves
the (default) timeout for `init/1` as `:infinity`.

Currently, in the spirit of the other error handling in this bit of
code, shapes that timeout are deleted rather than the entire shape cache
init failing. I'm not sure this is the right approach TBH -- some
critical problem in the storage layer or elsewhere could result in all
shapes being removed by this timeout and the system starting "normally"
but with some fundamental problem at the storage or pg layer.

Options are:

- as currently just drop shapes that timeout (with an error in the logs)
- allow some number of failures but raise after that
- crash with any timeout

Feedback welcome.

Fixes https://github.com/electric-sql/electric/issues/2574