fix(sync-service): Fix initialisation race-condition (#2576)
Fixes #2529 by introducing a StatusMonitor process that monitors:
- That the PG lock has been acquired
- The replication client is ready
- The Connection pool is running
- The shape log collector is ready 

This last condition was not previously waited for which could result in
the "the table identifier does not refer to an existing ETS table" error

The StatusMonitor also monitors all the relevant processes to check to
see if they have died and updates the status accordingly.

Thanks to @alco for suggesting this solution, and to @icehaunter for
recommending the awesome Repatch mocking library that this PR
introduces.

---------

Co-authored-by: Ilia Borovitinov <ilia@electric-sql.com>