fix: improve error handling around a periodic wal size check (#2604)
This fixes an error we're seeing sometimes in Sentry around monitoring.
Not sure about the root cause of process being unresponsive, but it
might be that it's a slow query against Postgres that's blocking the
entire thing. This PR moves the wal query into a task so that the
manager can continue to handle `DOWN` messages and such