Ensure installation ID inside config/runtime.exs (when possible)
The Erlang OTEL library wants to have all the resource attributes
available at the time of initialisation. Since it's a dependency, its
initialisation takes place before our Application.start() callback is
invoked.

I thought that writing a custom resource detector would circumvent this
problem by polling for the installation ID in the background, but it has
turned out that the resource detector runs until it succeeds or times
out __before__ Application.start() callback is invoked.

So the only way to set the installation_id on the OTEL resource is to
make sure it has been generated inside config/runtime.exs, before any
OTP applications are started. This also means we no longer need to poll
for it inside the resource detector.