fix: Electric restart shouldn't cause requests with `0_inf` to fail (#2624)
Issue was only with shapes that didn't observe any writes after being
created. When creating a snapshot, we don't know ahead of reading how
many total rows we would read from DB, so we created a shape with an
in-memory value for last known offset of `0_inf` - after all real
snapshot chunks (denoted as `0_1`, `0_2`, ..., `0_n`), but before real
LSNs with `x_y` where x > 0. On a restart however we'd read actually
created chunk lists and fill in "last known" offset to be say `0_2`
instead. Which meant suddenly clients who previously read through the
snapshot and were just hanging waiting for new changes got the 400
because requested `0_inf` was indeed larger than presumed maximum of
`0_2`. The PR just makes it so that the state of the world after a
restart is consistent with the one we achieve while creating a fresh
shape.

Closes #2596

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>