fix: Pause and resume shapestream on visibility changes (#2435)
https://github.com/electric-sql/electric/issues/2364 reported an issue
with Safari on iOS which caused the application to stop syncing changes.
The reason for this behavior is that Safari aggressively
throttles/suspends/cancels timers and loops to save battery. Therefore,
the loop in `ShapeStream.#start` was sometimes stopped and never
resumed.

I tried fixing this with timers (`setTimeout`/`setInterval`) but that
did not solve the issue. Instead, i introduced `pause` and `resume`
methods on the `ShapeStream`. When the TS client is running in the
browser, the `ShapeStream` now uses the [visibility
API](https://developer.mozilla.org/en-US/docs/Web/API/Page_Visibility_API)
to pause the stream when the app goes in the background and to resume it
when it comes back into the foreground.

I tested this fix using the React example with Safari on iOS.
To test it for yourself, this is the setup that was reproducing the
issue and works with this branch:
- Run ngrok to expose both `http://localhost:3000` (local PG) and
`http://localhost:5173` (local dev server)
- Run `pnpm backend:up`
- Run `VITE_ELECTRIC_URL="https://<your subdomain for PG>.ngrok.app"
pnpm vite`
- In Safari on iOS visit `https://<your subdomain for dev
server>.ngrok.app`
- Lock your iPhone
- Perform some update
- Unlock your iPhone, the app should now be in the old state and not
react to updates

## Gotchas

In order to grab the global `document`variable (when in browser
environments), i had to add the `DOM` and `DOM.Iterable` target
libraries to the TS config. However, because of that the type of the
`fetchClient` option for the `ShapeStream` changed since `typeof fetch`
was `string | URL` and is now `string | URL | Request`. This is a
breaking change since user-provided fetch functions will need to be
widened to also support a `Request` input argument.

Either we are fine with this or we remove those target libraries and
instead manually typecast the global `document` variable.

## Note to reviewers

Changes to `ShapeStream.#start` are small but i extracted the while loop
into an if test that returns fast:
```ts
while (
  (!this.options.signal?.aborted && !this.#isUpToDate) ||
  this.options.subscribe
) {
  <perform request>
}
```
Modified this to:
```ts
if (
  !this.options.subscribe &&
  (this.options.signal?.aborted || this.#isUpToDate)
) {
  return
}

<perform request>
```

Because of this change, the indentation of the code in `<perform
request>` changed and Git fails to detect this so it marks the original
code as deleted and the indented code as added.

---------

Co-authored-by: Kyle Mathews <mathews.kyle@gmail.com>