fix: handle publication/schema alterations on silent PG (#2634)
When PG is silent - there are no updates to tables that might interest
us - the we're not busting the cache of the schema, not invalidating
shapes for deleted tables, etc. This PR addresses that.

Practical impact:
- `PublicationManager` now tracks table OIDs along with the schema/name
pair.
- When altering the publication, we check for dropped, recreated, or
renamed tables. For any that match, we change the altering of the
publication to not touch the affected tables, and instead force a clear
on affected shapes. This is needed despite a periodic check, because
there is a window where this may come up between the checks.
- `PublicationManager` now has a bit of coupling to `ShapeCache` to
notify all the shapes that should be invalidated.
- For invalidated shapes, we're sending 409s, because maybe it's just a
table recreation and clients can start from scratch, or maybe it's a
table drop and further requests will yield a 400 when table is actually
no longer found.

Issue impact:
- Closes #2449 - decision here is that if the table has been recreated,
we shouldn't re-add it to the publication automatically, but rather
invalidate the shapes and creation of a new shape initiated by the
client might re-add the table if it need to.
- Closes #2558 - previous PR already closed this issue, but this PR
makes the window where the guard is necessary smaller by polling the
schema periodically
- Closes #1733 - addresses it in two ways, by periodic polling as well
as double-checking when actually doing the publication altering to have
a transactionally-consistent view
- Closes #2517 - schema changes on a silent DB are caught within a
polling period

---------

Co-authored-by: Rob A'Court <robacourt@users.noreply.github.com>