fix: Create and delete Electric sources for examples (#2646)
Closes https://github.com/electric-sql/electric/issues/2645

The optimal solution is to use a proper Pulumi provider which is in the
works, but to avoid blocking on that I've used a Pulumi command like
we've done for Neon.

It uses the output from the `create` command in the `delete` command in
order to delete the source using the ID generated in the creation step.

I've added a 10s sleep after the delete to ensure the Neon database is
not deleted until the Electric source is cleaned up - technically this
should be solved on the cloud (if cleanup fails cause the database no
longer exists we should consider the cleanup to be successful), but
adding it here to ensure cleanups go smoothly.

Only concern is if delete fails with 404 because source has already been
deleted it will currently not consider that "successful" and the stack
won't be removed entirely. Something we can live with I think until we
get the proper provider.

Tested this with local cloud dev stack.