Fix the overly optimistic transition to streaming in ReplicationClient
We used to assume that as soon as we instructed the ReplicationClient to
start streaming, it would transition into the streaming mode without
fail. That's not the case.

After the client issues the START_REPLICATION_SLOT statement to Postgres
and the Elixir process transactions into the streaming mode, it may
still exit with an error caused by a different PG backend already using
the slot.