fix: Add retries on shape file cleanup operation to prevent `EEXIST` race (#2662)
We observed the `unsafe_cleanup!` call on shape files failing with an
`:eexist` reason on the `File.rm_rf!` calls, which didn't make much
sense.

Looking at the Elixir source we suspect that the issue is `rm_dir`
failing due to some race occurring:

https://github.com/elixir-lang/elixir/blob/e35ffc5a903bff3b595e323eb1ac12c4ecd515ad/lib/elixir/lib/file.ex#L1308-L1312

Here's a discussion with Gemini about this as well
https://gemini.google.com/share/d36919a2b507

Potentially the directory is assumed to be empty after files have been
deleted but something causes it to not actually be fully empty,
especially on weirder filesystems like AWS' EFS which is networkd and
distributed.

This PR suggests adding some arbitrary retries with small waits in an
attempt to mitigate this issue - I'm open to suggestions about different
wait times (if any, potentially we just retry and since they are
syscalls they already will take some time) and retry attempts.

---------

Co-authored-by: Ilia Borovitinov <ilia@electric-sql.com>