feat(sync-service): Friendly error messages when failing to connect to the database (#2654)
Fixes #2276 . Expected errors now have helpful error messages with a
link to documentation where available.

Expected errors are either:
- Fatal in that they can't be fixed without user intervention. For stand
alone electric these errors get logged as a critical error and Electric
is shut down. For library/cloud mode a warning is logged and a
`:fatal_error` is publish to the stack registry.
- A retry may fix the issue, in which case a warning is logged and
recorded in the StatusMonitor so that if someone queries the API they
get a helpful message rather than "stack not ready". The error is also
published to the stack registry as a `:retryable_error`.

Unrecognised errors are logged as errors (and so Sentry will see them)
but not counted as fatal errors since retrying may fix the issue.

Current expected errors:
- Connecting issues:
  - domain not found
  - connection timeout
  - connection refused
- blocked by a firewall (this is either a timeout or connection refused)
- Postgres errors:
  - the username or password is incorrect
  - db doesn't exist
- user doesn't have REPLICATION role to setup the publication slot (we
return nice error about this & link to docs on creating a user with the
right role)
- the db doesn't have logical replication enabled — (again we return
nice error and link to docs on enabling logical replication)

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>