Refactor and fix connection initialization/reconnection backoff (#2652)
The highlight of this PR is the code refactoring inside
Connection.Manager module to make the state machine that it implements
more explicit and transitions between steps better documented.

Not too long ago we made the initialization of ReplicationClient
asynchronous which had been necessitated by the possibility of its
startup getting blocked on a pending DB transaction. This PR also makes
LockConnection's initialization async for consistency. Both types of
processes now return early from their respective `init()` callbacks and
then cast messages to the connection manager when reaching a new
milestone (connected, lock_acquired, ready_to_stream,
streaming_started).

What prompted this refactoring was a new failure state we had
discovered. If there's a pending transaction that blocks Electric's
initialization for the first time and Electric is then shut down,
Postgres will still have a running backend blocked on the
`CREATE_REPLICATION_SLOT` statement. When Electric starts again, it will
be able to complete the full connection setup (because the replication
slot has already been created in Postgres) but will fail when switching
ReplicationClient into streaming mode because it's aready actively used
by a different OS process (the `walsender` backend in Postgres stuck
from Electric's first launch).

Electric's failure to handle this specific scenario resulted in that it
got stuck in a reconnection loop without exponential backoff, always
trying to reconnect every 2 seconds. Additionally, it resulted in a
snowballing of warning logs with each new logger call getting added on
every reconnection attempt, due to the absence of any way to keep track
of periodic connection status checks and the logs they were emitting.

The PR fixes both the runaway logging and the error handling of this new
failure state. It is now simulated in the new integration test defined
in `replication-client-self-conflict.lux`.

A new unit test is added that shows the minimum required configuration
to start a Connection.Supervisor in isolation. Trimming it down further
to only start the Connection.Manager process would require some mocking
which I dislike.

As a nice bonus, I discovered a bug in reading the connection options
and pool options from application env: the app env wasn't actually being
consulted. Probably not a problem in Electric Cloud but it did result in
the `ELECTRIC_QUERY_DATABASE_URL` config to be ignored.