fix: Use regular connection for advisory lock instead of pooled one (#2682)
I'm investigating some issues around the release of the advisory lock -
sometimes I've noticed that we've terminated the Electric process but
when trying to grab the relevant advisory lock elsewhere we fail as the
lock is still in use.

I'm not sure how quickly Postgres should be responding to the session
ending and releasing the lock, I assume very quickly, but I started
thinking about what happens with connection pooling.

For example, pgBouncer [does not support session
locks](https://www.pgbouncer.org/features.html) when in transaction
pooling mode, and generally unless advisory locks are explicitly
supported poolers would mess with the expected behaviour of the advisory
lock.

For that reason I suggest using the unpooled connection for the advisory
lock. I also thought that we might want to explicitly unlock it but
according to PG docs:
> pg_advisory_unlock_all () → void : Releases all session-level advisory
locks held by the current session. (This function is implicitly invoked
at session end, even if the client disconnects ungracefully.)

---------

Co-authored-by: Oleksii Sholik <oleksii@sholik.dev>