fix: Only log unexpected lock acquisition errors (#2684)
The lock acquisition might take a while, and many PG providers might
time out the statement execution - we're fine to just keep trying but we
should only log unexpected errors.

We could argue that we don't even want to backoff and retry, and just
immediately retry, but I'm worried that since the providers are timing
statements out we might be better off not retrying immediately.