fix: Add timeout before considering replication connection successful (#2693)
Closes https://github.com/electric-sql/electric/issues/2691

Added a 5 second timeout - what we've observed is setup errors would
come through within a few milliseconds, but I think allowing a few
seconds is reasonable. A connection that falls over within 5 seconds of
being established should be considered unsuccessful IMO.

I've avoided using timers and storing timer references, as I think we
can get the job done by relying on the fact that if the replication
client pid is the same between receiving the first message and the
timeout passing, we can consider the connection successful.

Will test this ad hoc with @alco 's reproduction steps - we don't really
have proper testing around the ConnMan so not sure what to do with this.