fix: Remove location header from 409 responses (#2713)
Closes https://github.com/electric-sql/electric/issues/2504

Technically we were documenting this in our OpenAPI spec, so even though
we don't use it in our clients and we don't advertise it much in our
docs this _is_ a bit of a breaking change?

I'll let @KyleAMathews decide if we want to move forward with this, I've
prepared this PR to get rid of it just to accelerate a decision on this.