chore (examples): Use a shared PG for production deployment of examples (#2726)
This PR is a first step towards using a shared PG DB for all production
deployments of the examples.
Currently, it only uses this shared DB for the React example. After
verifying that this works, we can open a follow up PR to use this shared
DB for all examples that use the items table.