sync-service: Use separate mock modules per-test (#2740)
keeping the same Mock.* naming

This prevents expectations defined in simultaneously executing tests
from clashing. I think this is the cause of various flakes in the CI -
we define expectations on `Support.Mock.Behaviour` and occasionally a
parallel test hits that expectation but fails because it's coming from
an unregistered pid.

This pr gives every test module its own set of
`__MODULE__.Mock.Behaviour` mocks to define expectations on but keeps
the same usage in the tests.