chore (examples): Add playwright tests to common examples (#2738)
This PR adds playwright tests to all of the examples.
Most examples have SST code to auto deploy them on PRs and to production
(when merged in main).
For those examples, we also run the playwright tests in PRs.

Some examples don't have SST code and were deployed manually.
For those, we only run the playwright tests periodically on the
production deploy of those examples.

Currently, we run playwright tests on the changed examples in PRs and in
production (when merged to main).
We also test all examples once a day. **Should we also test all examples
after a new release** (since that triggers a cloud upgrade)**?** Note
that we can also manually trigger the workflow that tests all examples.

Also fixes https://github.com/electric-sql/electric/issues/2752