fix(sync-service): Increase restore timeout (#2750)
Fixes #2741 . The [20s timeout was in version
1.0.3](https://github.com/electric-sql/electric/blob/8c0a22aaf676552ebc38efc7e03481c6c1b6db37/packages/sync-service/lib/electric/shape_cache.ex#L96)
which trigger.dev are using, however we've replace the timeout for a
shape specific timeout in the latest version which has mostly sorted the
issue, and probably would fix trigger.dev.

Also, since 1.0.3 we now wait for the shape log collector to be ready
before calling it, so the `create_or_wait_shape_handle` timeouts we see
in trigger.dev's logs won't happen either.

There is still one place it can time out which is the `start_processing`
call, since this waits until all consumers have subscribed and
surprisingly this can happen _after_ the consumers are all restored, so
this PR increases the timeout of that to 60s. I haven't made it scale
with the number of shapes or be configurable as yet as as long as
subscription time is roughly the same as restoration time which it
typically is, no time is used at all, if we find this timeout is
actually hit we'll be in a better position to see what the best solution
is.